+++ packages/sdk/src/constants/CicsCmci.constants.ts
@@ -19,6 +19,11 @@ export const CicsCmciConstants = {
   CICS_SYSTEM_MANAGEMENT: "CICSSystemManagement",
 
   /**
+     * Specifies the required part of the REST interface URI to access system initialization parameters
+     */
+  CICS_SYSTEM_PARAMETER: "CICSSystemParameter",
+
+  /**
      * Specifies the required part of the REST interface URI to access program definitions
      */
   CICS_DEFINITION_PROGRAM: "CICSDefinitionProgram",
@@ -39,6 +44,11 @@ export const CicsCmciConstants = {
   CICS_PROGRAM_RESOURCE: "CICSProgram",
 
   /**
+     * Specifies the required part of the REST interface URI to access library resources
+     */
+  CICS_LIBRARY_RESOURCE: "CICSLibrary",
+
+  /**
      * Specifies the required part of the REST interface URI to access URIMap definitions
      */
   CICS_DEFINITION_URIMAP: "CICSDefinitionURIMap",
@@ -47,6 +57,11 @@ export const CicsCmciConstants = {
      * Specifies the required part of the REST interface URI to access webservice definitions
      */
   CICS_DEFINITION_WEBSERVICE: "CICSDefinitionWebService",
+
+  /**
+     * Specifies the required part of the REST interface URI to access tcp/ip service resources
+     */
+  CICS_TCPIPSERVICE_RESOURCE: "CICSTCPIPService",
 
   /*
      * Specifies the required part of the REST interface URI to access URIMaps