+++ packages/vsce/src/trees/CICSCombinedTrees/CICSCombinedTCPIPServiceTree.ts
@@ -20,6 +20,7 @@ import { TextTreeItem } from "../treeItems/utils/TextTreeItem";
 import { getFolderIcon } from "../../utils/iconUtils";
 import { ViewMore } from "../treeItems/utils/ViewMore";
 import { CICSTCPIPServiceTreeItem } from "../treeItems/web/treeItems/CICSTCPIPServiceTreeItem";
+import { CicsCmciConstants } from "@zowe/cics-for-zowe-sdk";
 
 export class CICSCombinedTCPIPServiceTree extends TreeItem {
   children: (CICSTCPIPServiceTreeItem | ViewMore)[] | [TextTreeItem] | null;
@@ -40,7 +41,8 @@ export class CICSCombinedTCPIPServiceTree extends TreeItem {
     this.activeFilter = undefined;
     this.currentCount = 0;
     this.incrementCount = +`${workspace.getConfiguration().get("zowe.cics.allTCPIPS.recordCountIncrement")}`;
-    this.constant = "CICSTCPIPService";
+    this.constant = CicsCmciConstants.CICS_TCPIPSERVICE_RESOURCE;
+
   }
 
   public async loadContents(tree: CICSTree) {
@@ -60,6 +62,7 @@ export class CICSCombinedTCPIPServiceTree extends TreeItem {
           let count;
           const cacheTokenInfo = await ProfileManagement.generateCacheToken(
             this.parentPlex.getProfile(),
+            this.getSession(),
             this.parentPlex.getPlexName(),
             this.constant,
             criteria,
@@ -72,6 +75,7 @@ export class CICSCombinedTCPIPServiceTree extends TreeItem {
               if (recordsCount <= this.incrementCount) {
                 allTCPIPS = await ProfileManagement.getCachedResources(
                   this.parentPlex.getProfile(),
+                  this.getSession(),
                   cacheTokenInfo.cacheToken,
                   this.constant,
                   1,
@@ -80,6 +84,7 @@ export class CICSCombinedTCPIPServiceTree extends TreeItem {
               } else {
                 allTCPIPS = await ProfileManagement.getCachedResources(
                   this.parentPlex.getProfile(),
+                  this.getSession(),
                   cacheTokenInfo.cacheToken,
                   this.constant,
                   1,
@@ -151,6 +156,7 @@ export class CICSCombinedTCPIPServiceTree extends TreeItem {
         }
         const cacheTokenInfo = await ProfileManagement.generateCacheToken(
           this.parentPlex.getProfile(),
+          this.getSession(),
           this.parentPlex.getPlexName(),
           this.constant,
           criteria,
@@ -162,6 +168,7 @@ export class CICSCombinedTCPIPServiceTree extends TreeItem {
           const count = recordsCount;
           const allTCPIPS = await ProfileManagement.getCachedResources(
             this.parentPlex.getProfile(),
+            this.getSession(),
             cacheTokenInfo.cacheToken,
             this.constant,
             this.currentCount + 1,
@@ -205,5 +212,9 @@ export class CICSCombinedTCPIPServiceTree extends TreeItem {
 
   public getParent() {
     return this.parentPlex;
+  }
+
+  public getSession() {
+    return this.getParent().getSession();
   }
 }